Many modifications (read description). Program doesn't compile.
First of all, the program is not working, and I can't imagine why.

I have created a strcmpi function in String.h
'>>' and '<<' operators implemented for vector class. The program compiled before, although there is some error when we try to input a vector.
Vector3D has got '<<' and '>>' operators as well for file and stream input and output.