Problems with << and >> corrected, read description (interesting stuff)
I was getting a problem with the call
outFile << tester;
Where outFile is an ofstream and tester is a vector<int>.
The problem was that Vector3D had a constructor Vector3D( x=0, y=0, z=0), where the zero values are default. Also, ofstream << vector was implemented as to operate << into each element of vector. But, as 'tester' is a vector of int's, when the compiler tried to do outFile << tester[0], it just didn't know if it was to print tester[0] or to initialize a Vector3D with tester[0] as a parameter and then print.
It is problematic to have something like Class(Type x), because, everytime we call function(x) where 'function' is overloaded to work both with Class and Type, the compiler won't know if we do mean function(x) or function(Class(x)). To correct this, the 'explicit' keyword is needed, so 'x' means only 'x'. If we want Class(x), we need to write Class(x) explicitly.